+++ NEWS.md
@@ -8,6 +8,72 @@ Cite all releases with doi [10.5281/zenodo.8226419](https://doi.org/10.5281/zeno
 which always resolves to the latest release.
 
 
+# PFUPipelineTools 0.1.13 (2025-05-15)
+
+* `pl_upsert()` gains new arguments `round_double_columns` and `digits` 
+  and calls `round_double_cols()` if requested.
+* `round_double_cols()` rounds double-precision columns 
+  in a data frame to assist the `compress()` function 
+  in the database.
+* `pl_upsert()` gains new argument `compress`, 
+  which defaults to `FALSE`.
+  If set to `TRUE`, `compress_rows()` is called internally
+  within `pl_upsert()`.
+* New functions `install_compress_function()`, 
+  `remove_compress_function()`, and 
+  `compress_rows()` 
+  assist with compressing rows in the remote database.
+* `pl_filter_collect()` gains new argument `...`
+  in which filtering expressions can be placed.
+  `...` replaces the myriad other arguments that
+  provided filtering for possible columns. 
+  This is a breaking change, but the 
+  new approach provides significant
+  flexibility for users of this function.
+  Plus, the code is much cleaner inside `pl_filter_collect()`!
+* In several places, 
+  code no longer passes `conn`
+  when both `schema` and `fk_parent_tables` 
+  are known.
+* Added a safety check in `pl_destroy()` to 
+  disable destroying "MexerDB".
+* `pl_collect_from_hash()` and `pl_filter_collect()`
+  gain new argument `version_string`
+  that provides capability to filter downloads by version.
+  The default value (`NULL`) downloads all versions.
+  Multiple versions can be downloaded by passing
+  a vector of strings.
+  Supplying `c()` (an empty vector)
+  downloads a table with no rows.
+* `pl_collect_from_hash()` is now more convenient
+  with default arguments for `schema` and `fk_parent_tables`
+  that pull values from `conn`.
+  This new behavior for `pl_collect_from_hash()` 
+  is now consistent with `pl_filter_collect()`.
+* New tests for new features. 
+    - Now up to 233 tests, all passing
+    - Test coverage now reported to be 25.27%, but 
+      that's an undercount. 
+      Many (most?) functions are not tested 
+      on continuous integration platforms and CRAN. 
+
+
+# PFUPipelineTools 0.1.12 (2024-12-09) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.14589472.svg)](https://doi.org/10.5281/zenodo.14589472)
+
+* Improved defaults for arguments to `pl_filter_collect()`.
+* No new tests. 
+    - Still at 143 tests, all passing
+    - Test coverage now reported to be 27.37%
+
+
+# PFUPipelineTools 0.1.11 (2024-12-09)
+
+* Updates to many accessing functions for the database.
+* Added several new tests for new accessing functions.
+    - Now up to 143 tests, all passing
+    - Test coverage now at 92.73%
+
+
 # PFUPipelineTools 0.1.10 (2024-08-01)
 
 * Eliminated the `IEAMW` column everywhere.